backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vasilii Kletkin <kletkinvasilii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "<5.0"
telethon = "^1.34.0"
django-telethon = "^1.2.4"
celery = "^5.3.6"
django-model-utils = "^4.4.0"
redis = "^5.0.2"
django-extensions = "^3.2.3"
python-socks = "^2.4.4"
django-celery-results = "^2.5.1"
requests = "^2.31.0"
django-celery-beat = "^2.5.0"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
proxy-checker = "^0.6"
pysocks = "^1.7.1"
django-taggit = "^5.0.1"
django-autocomplete-light = "^3.11.0"
django-debug-toolbar = "^4.3.0"
dal-admin-filters = "^1.1.0"
django-admin-rangefilter = "^0.12.1"
debugpy = "^1.8.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
